requirements.txt
anthropic>=0.40.0
python-telegram-bot>=20.7
playwright>=1.45.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
Pillow>=10.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
aiohttp>=3.9.0
requests>=2.31.0
